Extract `createInclusionFilter` to dir module
I intend to reuse this for a non-copy purpose.
